Работаю над уменьшением связанности кода
