+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,0 +1,17 @@
+//
+//  QuizResultsViewModel.swift
+//  MovieQuiz
+//
+//  Created by Vladislav Mishukov on 20.07.2023.
+//
+
+import UIKit
+
+struct QuizResultsViewModel {
+    // строка с заголовком алерта
+    let title: String
+    // строка с текстом о количестве набранных очков
+    let text: String
+    // текст для кнопки алерта
+    let buttonText: String
+}
